@@ -74,14 +74,14 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
+    "@babel/core": "^7.12.17",
+    "@babel/eslint-parser": "^7.12.17",
     "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/preset-env": "^7.12.16",
+    "@babel/preset-env": "^7.12.17",
     "@babel/preset-react": "^7.12.13",
     "@mdx-js/loader": "^1.6.21",
     "@octokit/auth-action": "^1.3.2",
-    "@octokit/rest": "^18.1.0",
+    "@octokit/rest": "^18.2.0",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.4.2",
     "@svgr/webpack": "^5.5.0",
     "alex": "^5.1.0",
@@ -90,28 +90,29 @@
     "copy-webpack-plugin": "^7.0.0",
     "css-loader": "^5.0.2",
     "css-minimizer-webpack-plugin": "^1.2.0",
-    "cypress": "^6.4.0",
+    "cypress": "^6.5.0",
     "directory-tree": "^2.2.6",
     "directory-tree-webpack-plugin": "^1.0.2",
     "docschina-remark-slugger": "^0.1.4",
     "duplexer": "^0.1.1",
-    "eslint": "^7.19.0",
+    "eslint": "^7.20.0",
     "eslint-config-prettier": "^7.2.0",
     "eslint-plugin-cypress": "^2.11.2",
-    "eslint-plugin-markdown": "^2.0.0-rc.2",
+    "eslint-plugin-markdown": "^2.0.0",
     "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "front-matter": "^4.0.2",
-    "html-loader": "^1.3.0",
-    "html-webpack-plugin": "^5.0.0",
+    "html-loader": "^2.0.0",
+    "html-webpack-plugin": "^5.1.0",
     "http-server": "^0.12.3",
     "husky": "^5.0.9",
     "hyperlink": "^4.6.0",
     "jest": "^26.6.3",
     "lint-staged": "^10.5.4",
     "lodash": "^4.17.19",
-    "markdownlint": "^0.23.0",
+    "markdownlint": "^0.23.1",
     "markdownlint-cli": "^0.26.0",
-    "mini-css-extract-plugin": "^1.3.6",
+    "mini-css-extract-plugin": "^1.3.8",
     "mkdirp": "^1.0.4",
     "modularscale-sass": "^3.0.3",
     "node-fetch": "^2.6.1",
@@ -132,7 +133,7 @@
     "remark-refractor": "git://github.com/montogeek/remark-refractor.git",
     "remark-slug": "^6.0.0",
     "rimraf": "^3.0.2",
-    "sass": "^1.32.7",
+    "sass": "^1.32.8",
     "sass-loader": "^11.0.1",
     "sirv-cli": "^1.0.11",
     "sitemap-static": "^0.4.2",
@@ -143,7 +144,7 @@
     "textlint": "^11.6.3",
     "textlint-rule-heading": "^1.0.10",
     "unist-util-visit": "^2.0.3",
-    "webpack": "^5.21.2",
+    "webpack": "^5.23.0",
     "webpack-bundle-analyzer": "^4.4.0",
     "webpack-cli": "^4.5.0",
     "webpack-dev-server": "^4.0.0-beta.0",
@@ -171,6 +172,7 @@
   },
   "resolutions": {
     "sitemap-static/minimist": "1.2.5",
-    "ini": "1.3.7"
+    "ini": "1.3.7",
+    "eval": "^0.1.5"
   }
 }
